Move SVGStyleParser tests out of other classes
